+++ app/Http/Controllers/Admin/DataSetController.php
@@ -570,9 +570,12 @@ class DataSetController extends AdminController
                     'action'        => 'create',
                 ], $response['data']));
             } else {
-                // delete resource record on fail
+                // Delete resource record on fail
                 $failMetadata = Resource::where('uri', $response['uri'])->forceDelete();
-                $request->session()->flash('alert-danger', __('custom.changes_success_fail'));
+                $request->session()->flash(
+                    'alert-danger',
+                    empty($response['data']['error']) ? __('custom.changes_success_fail') : $response['data']['error']
+                );
 
                 return redirect()->back()->withInput()->withErrors($response['errors']);
             }
@@ -821,7 +824,10 @@ class DataSetController extends AdminController
                         'action'        => 'update',
                     ], $response['data']));
                 } else {
-                    $request->session()->flash('alert-danger', __('custom.changes_success_fail'));
+                    $request->session()->flash(
+                        'alert-danger',
+                        empty($response['data']['error']) ? __('custom.changes_success_fail') : $response['data']['error']
+                    );
 
                     return redirect()->back()->withInput()->withErrors($response['errors']);
                 }