+++ resources/views/user/orgDatasetEdit.blade.php
@@ -9,7 +9,9 @@
         @include('partials.user-nav-bar', ['view' => 'organisation'])
     @endif
     @include('partials.org-nav-bar', ['view' => 'dataset', 'organisation' => $fromOrg])
-    @include('partials.org-info', ['organisation' => $fromOrg])
+    <div class="col-sm-3 col-xs-12">
+        @include('partials.org-info', ['organisation' => $fromOrg])
+    </div>
     @include('components.datasets.dataset_edit', ['admin' => Auth::user()->is_admin])
 </div>
 @endsection